Rename ItemForm-related `marketValueUSD` to `coverageAmountUSD`
This brings us in line with the property names returned by the API for
PolicyItems. When dealing with a PolicyItem, there is a coverage_amount
field but no market_value field. This change updates item-related code
to use the appropriate name.

Fair market value is a concept used when dealing with a claim, not when
(otherwise) dealing just with an item.